fused sending waiting times + blacklist into a single request (iallgather)

+++ ExaHyPE/exahype/stealing/AggressiveCCPDistributor.cpp
@@ -216,9 +216,9 @@ void exahype::stealing::AggressiveCCPDistributor::updateLoadDistribution() {
   int nnodes = tarch::parallel::Node::getInstance().getNumberOfNodes();
   int myRank = tarch::parallel::Node::getInstance().getRank();
 
-  int *waitingTimesSnapshot = new int[nnodes*nnodes];
+  double *waitingTimesSnapshot = new double[nnodes*nnodes];
  // const int* currentWaitingTimesSnapshot = exahype::stealing::PerformanceMonitor::getInstance().getWaitingTimesSnapshot();
-  const int* currentWaitingTimesSnapshot = exahype::stealing::StealingAnalyser::getInstance().getFilteredWaitingTimesSnapshot();
+  const double* currentWaitingTimesSnapshot = exahype::stealing::StealingAnalyser::getInstance().getFilteredWaitingTimesSnapshot();
   std::copy(&currentWaitingTimesSnapshot[0], &currentWaitingTimesSnapshot[nnodes*nnodes], waitingTimesSnapshot);
 
   //waitingTimesSnapshot[myRank] = waitingTime;
